requirements.txt
google-api-python-client
python-dotenv
streamlit
aws-wsgi
transformers==4.42.4
tensorflow==2.17.0
pandas
datasets
tf-keras
textblob
emoji
sentencepiece 
googletrans==4.0.0-rc1